backend/lambda/requirements.txt
boto3==1.35.0
psycopg2-binary==2.9.9
cryptography==43.0.0
